feat: add job preview page
