+++ src/data/jobs.js
@@ -5,7 +5,6 @@ const jobs = [
     description:
       'ด้วยลักษณะนิสัยที่มีความรอบคอบ มีเหตุและผล ทำสิ่งใดโดยคาดหวังผลลัพธ์ที่มีประสิทธิภาพ ภายในกรอบและเวลาที่กำหนด เป็นคนที่พึ่งพาได้ และมีแนวคิดตามหลักการผิดถูกชัดเจน',
     img: '',
-    to: '#',
   },
   {
     job: 'Quality Assurance',
@@ -13,7 +12,6 @@ const jobs = [
     description:
       'ด้วยลักษณะนิสัยที่ละเอียดรอบคอบ ทำงานโดยเน้นประสิทธิภาพของกระบวนการการทำงาน และในขณะเดียวกันก็ปฏิบัติตามกฏและหลักการอย่างเคร่งครัด',
     img: '',
-    to: '#',
   },
   {
     job: 'Support Engineer',
@@ -21,7 +19,6 @@ const jobs = [
     description:
       'คุณมองผลลัพธ์เป็นที่ตั้ง ตัดสินใจและลงมือทำงานตามข้อกำหนดและคำแนะนำที่ถูกที่ควร เพื่อแก้ปัญหาให้ได้อย่างรวดเร็วที่สุด สามารถเป็นที่ปรึกษาที่ดีและให้คำตอบหรือชี้นำทางออกที่เหมาะสมได้',
     img: '',
-    to: '#',
   },
   {
     job: 'Scrum Master',
@@ -29,7 +26,6 @@ const jobs = [
     description:
       'คุณเป็นคนที่มีรูปแบบการทำงานที่เป็นระบบชัดเจน แต่ก็ใส่ใจกับเพื่อนร่วมงานและคนรอบข้าง มีความเข้าใจและพร้อมที่จะยืดหยุ่นตามสถานการณ์ที่เหมาะสม แต่ก็ยังยึดตามขอบเขตของงานและระยะเวลาที่กำหนด',
     img: '',
-    to: '#',
   },
   {
     job: 'Product Owner',
@@ -37,7 +33,6 @@ const jobs = [
     description:
       'ด้วยลักษณะนิสัยที่ชอบหาข้อมูล ความเป็นไปได้ และเปรียบเทียบตัวเลือกก่อนตัดสินใจ สามารถพลิกแพลงวิธีการแก้ปัญหาให้เหมาะสมกับกรอบเวลา มีความยืดหยุ่นในการทำงาน และสามารถจัดลำดับความสำคัญได้ดี',
     img: '',
-    to: '#',
   },
   {
     job: 'Account Manager',
@@ -45,7 +40,6 @@ const jobs = [
     description:
       'คุณเป็นคนที่ความสำคัญกับทั้งวิธีการและผลลัพธ์ของการทำงาน แม้จะทำงานภายใต้กรอบและข้อกำหนดที่ตั้งเอาไว้ ก็ยังมีลักษณะนิสัยที่มีความคิดริเริ่ม และเป็นผู้นำ สามารถให้คำแนะนำกับผู้อื่นได้ดี',
     img: '',
-    to: '#',
   },
   {
     job: 'People Manager',
@@ -53,13 +47,11 @@ const jobs = [
     description:
       'ด้วยลักษณะนิสัยที่ละเอียดรอบคอบ แต่ทว่ายืดหยุ่นและพลิกแพลงไปตามสถานการณ์ สามารถจัดลำดับความสำคัญได้ดี มีความกระตือรือร้น เข้ากับคนง่าย ให้ความสำคัญกับทั้งวิธีการและความราบรื่นในการทำงานของทีม',
     img: '',
-    to: '#',
   },
   {
     job: 'No match 🥲',
     description: '',
     img: '',
-    to: '#',
   },
 ];
