+++ src/data/label.js
@@ -3,4 +3,6 @@ export const label = {
   startButton: 'เริ่มเลย',
   questionHeading: 'เลือกคำตอบที่ใช่ตัวคุณที่สุด',
   jobIs: 'จากคำตอบของคุณ ลักษณะงานที่เหมาะสมคือ',
+  share: 'Share',
+  tryQuestionaire: 'ลองทำแบบสอบถาม',
 };
