+++ src/pages/job/[jid]/preview.js
@@ -1,0 +1,5 @@
+import Job from '@/components/Job';
+
+export default function JobPreviewPage() {
+  return <Job mode="preview"></Job>;
+}
